Move FileService tests to corresponding org.codefreak.codefreak.service.file package
